split ppm header class
