+++ README.md
@@ -43,7 +43,11 @@ P6
 IPpmImage ppmImage = await PpmImageReader.ReadAsync(@"C:\image.ppm");
 
 // Read raw image
-IImage rawImage = await RawImageReader.ReadAsync((@"C:\image.raw", 1920, 1080, 12, 0);
+(int width, int height, int pixelBits) = (640, 480, 12);
+IImage rawImage1 = await RawImageReader.ReadAsync(@"C:\image1.raw", width, height, pixelBits);
+
+(int stride, int pixelOffset) = (643, 0x10);
+IImage rawImage2 = await RawImageReader.ReadAsync(@"C:\image2.raw", width, height, pixelBits, stride, pixelOffset);
 ```
